Remove banned password check job
We're just going to check this at login time and deal with it then.
Still to be designed, but one idea that's been surfaced is to display
a message saying that their password is weak, and to lock the account
if the user don't change their password by the end of the session, or
in N days time.