Remove detect-secrets
We installed this pre-commit hook to help prevent us from leaking
production secrets, but the benefit hasn't outweighed the pain.  I
think we already solve this problem with good practices:

1. We don't share secrets between production and development/test
envs.

2. We specify our production secrets in environment variables.

So we've never been in the habit of committing real secrets to the
repo.  It seems strange that we'd slip up and do so now.

Furthermore, as this is only a pre-commit hook, and not a pre-receive
hook, it doesn't actually prevent someone from pushing a secret to
GitHub!  Since we have the same vulnerability whether have this hook
or not, I don't think the potential gain (someone gets access to a
production secret, hard-codes it, and then has that commit rejected
before pushing it) is worth the effort (needing to keep the baseline
and pragma comments up to date).