Install bullet gem to detect certain activerecord problems
Bullet is able to detect:

- N+1 query problems
- Eager-loaded associations which are unused
- COUNT queries which could be avoided with a counter_cache

We don't currently have any of these problems, but we easily could do!

For example, if we were to change the security page to show the user
agent of each security activity,

    +          <br>
    +          <%= activity.user_agent&.name %>

Then we would have an N+1 query problem if we didn't also eager-load
the `user_agent` association, and indeed bullet detects this, throws
an exception, and tells you what to change:

     Bullet::Notification::UnoptimizedQueryError:
       user: root
       GET /account/security
       USE eager loading detected
         SecurityActivity => [:user_agent]
         Add to your query: .includes([:user_agent])
       Call stack
         /govuk/govuk-account-manager-prototype/app/views/security/show.html.erb:84:in `block in _app_views_security_show_html_erb__3613801695823746490_88320'
         /govuk/govuk-account-manager-prototype/app/views/security/show.html.erb:70:in `each'
         /govuk/govuk-account-manager-prototype/app/views/security/show.html.erb:70:in `_app_views_security_show_html_erb__3613801695823746490_88320'
         /govuk/govuk-account-manager-prototype/spec/requests/security_activity_spec.rb:198:in `expect_event_on_security_page'
         /govuk/govuk-account-manager-prototype/spec/requests/security_activity_spec.rb:31:in `block (2 levels) in <top (required)>'

Even though it's not found any problems in the app so far, I think
it's a useful future-proofing thing to bring in.