Add basic grafana annotations for deploys
A simple annotation to show the start and end of a deployment event. Our
grafana dashboards tend to show spikes around deploy times.

Annotations should help make it easier to see if deploy events had any
impact on our traffic and also consider which events we should account
for when planning autoscaling strategies.