Format phone number when saving RegistrationState
We are currently sending a hotchpotch of differently formatted phone
numbers to Notify when users first register. This is because the
formatting to E.164 format only occurs once the User object is created.

Therefore we should also format the number during the registration, to
ensure the correctly formatted number is sent to Notify.