Scale down to 5 worker instances
These workers are mostly used for sending emails and text messages,
for running scheduled database cleanup tasks, and once a day one of
them spends a couple of minutes generating the BigQuery report.

All of these things are quick.  The workers are idle most of the time.
We don't need 20 of them, so I divided the count by 4 somewhat
arbitrarily.