Migrate ApplicationKey to a non-composite_primary_keys table
We're not getting much use out of this gem, and it's not compatible
with rails 6.1 yet (admittedly, they are working on it).  As it is a
surprisingly complex gem for little gain I think it's more of a
liability than a help, at least for our use-case where we only have a
single table with a composite primary key.

This will not be an error-free migration: any JWTs which are
validated *after* the migration but *before* the code deploy will
error; but I'm not sure there is a nice way to do this with
zero-downtime, and deploys are much faster now that we've scaled down
anyway.