Enable redis cache in production
The PaaS Redis instances are configured, by default, with the
`volatile-lru` eviction policy[1].  When the memory is full, the
least-recently-used key with an expiration time is purged.  So it's
safe to use our existing Redis as a cache, as well as for Sidekiq, so
long as we set a default expiration time.

Currently we don't have a cache configured, which I think means each
instance has its own in-memory cache capped to 32MiB - which makes the
IP address caching in SecurityActivity almost pointless.

[1] https://docs.cloud.service.gov.uk/deploying_services/redis/#redis-key-eviction-policy