Amend page titles and add meta descriptions
Page titles on the account manager are currently either blank
(defaulting to "GOVUK Account") or are using the page heading as a page
title. There are cases where using the page heading as the page title is
not appropriate.
There has been a content audit of all the page titles on the GOV.UK
Account journey. This implements the changes suggested in the content
audit.

Additionally, this adds meta descriptions to the pages that need them.
This will make GOVUK Account pages look more legitimate in search engine
results, or if shared on social media.