Update msgpack to version 1.4.2
The test image build for the GOV.UK Account manager has failed:
https://cd.gds-reliability.engineering/teams/govuk-tools/pipelines/govuk-account-manager-prototype/jobs/build-tests-image/builds/381

This is the message we're currently seeing:
```
Your bundle is locked to msgpack (1.4.1), but that version could
not be found in any of the sources listed in your Gemfile. If you
haven't changed sources, that means the author of msgpack (1.4.1) has
removed it. You'll need to update your bundle to a version other than
msgpack (1.4.1) that hasn't been removed in order to install.
```

Since msgpack 1.4.2 is available in Rubygems
(https://rubygems.org/gems/msgpack/versions/1.4.2), updating to this
version should solve the problem.