Conditionally route to signup with JWT
The header has an anchor-tag in the product name / crown that links to
the application root path (`/`). Before this commit, users who clicked
this link were routed to the signup page. This is not currently
desirable because we only want users to reach this page if they're
coming from the Brexit checker. This is because signing up without the
Brexit checker data results in an empty account, which is not useful for
the user.

We initially hoped we could redirect the user based on the href in the
account header. However, we use a [component][1] that has a hard coded
[href][2] so we cannot add conditionality at the view layer.

We then looked at doing this at the controller level. Visitors to the
root path are sent to the WelcomeController. This controller parses a
JWT and checks if the user is logged in. The new logic only applies if
the user is not logged in. If the user is logged in, the flow proceeds
as it did before.

Since JWT objects are only used in the context of the signup form, we
added a conditional statement to check if a JWT can be found with a
payload. If so, we progress to signup, otherwise we redirect the user to
login.

This works irrespective of if the JWT has arrived as a HTTP parameter,
or if it has been found by an id stored in the session.

[1]: https://github.com/alphagov/govuk-account-manager-prototype/blob/d49d5fa449029549ba75a967f7a5755dc72339e9/app/views/layouts/application.html.erb#L58-L62
[2]: https://github.com/alphagov/govuk_publishing_components/blob/19cd9a8e08563039f4e3e57a0a105884aa2be557/app/views/govuk_publishing_components/components/layout_header/_header_logo.html.erb#L2

Trello card: https://trello.com/c/mbhGiEIV/592-update-the-govuk-account-link-to-take-users-to-the-sign-in-journey

Co-authored-by: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>
Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>