Create webauthn credentials table
The app will store a multiple webauthn credentails, identified by an
external or rawId (the same ID a browser will use to look up a
credentail in it's own store).

We will also store a Public Key, which we can use to confirm the device
has cryptographically proven they can sign data with their private key.

We also store some other niceties, like a nickname, so the user knows
which credentail is which, and a sign count which can be used to track
how many times they have sucessfully signed data with this key

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>