Show an interstitial page for users with insecure passwords
This only shows up for users logging straight in to their account,
rather than going through an OAuth flow.  This is because I didn't
want to pass the `previous_url` logic through to the change password
form.

To make it easy to tell if the user is just logging in or is
OAuth-ing, I made the `LoginState` have `redirect_path: nil` in the
former case.