Add Acount deletion security event
This will allow us to check for suspicious activity in the event of a
sucessful user deletion request.

Note: the tests are a bit strange. Delete calls a dependent destroy on
SecurityActivities stored in the local database, so we cannot test for a
persisted event at the end of the test.

The real output is piped to STDOUT where it is picked up by splunk, so
in this case we can test for a :debug level event containing the action
we know will be set from the security activity