Preferentially take the JWT ID from the OAuth state param
We're retiring the POST method of submitting attributes to create when
a user signs up.  This is so we can send users on a normal OAuth login
journey.

This change is the first step towards that goal.  Next we'll be adding
an endpoint for an app to submit a JWT and get the ID to pass in the
state param.

Ultimately we'll be retiring the ability to set attributes on
registration too when the Transition Checker goes away, at which point
this whole JWT thing can go away.