Add an endpoint to accept a JWT and return its ID
Since the JWT comes from an authenticated endpoint, we don't need to
validate the crypto.  The only way to get an access token is to know
the client's secrets, and so if an attacker can explot this new way,
they could already exploit the old way.

The new endpoint rejects user access tokens, which ensures that only
someone able to do a client_credentials auth flow can create a JWT.

The Transition Checker JWT flow can be changed to:

1. [frontend] POST to a new endpoint in the Transition Checker
2. [backend] Generate an access token with the client credentials flow
3. [backend] POST the JWT to the account-manager, returning an ID
4. [backend] Generate an OAuth redirect URL with `state={returned id}`
5. [backend] Redirect the user to do a normal OAuth consent journey

This is good because we're dropping the weird POST-to-the-auth-server
flow we currently have, and adopting something more standard.  There
is a follow-up card to make the change to finder-frontend and to
delete the then-redundant POST code from this app.

The post_login_oauth field isn't needed any more because the standard
OAuth auth flow passes along the redirect URL.