Send a user to register, if they have a JWT, from doorkeeper
Currently we send the user from GOV.UK to the accounts domain in two
distinct ways:

- A link in the header.  This is a GET request and should go to the
sign in page.

- A form on the results page.  This is a POST request and should go to
the sign up page.

Currently this works because the link goes to the OAuth auth endpoint
and the form goes to the welcome route.  But when we switch to using
GET-style JWTs, the form will also go to the OAuth auth endpoint.

So we need to make the routing decision either in the session
controller, or in Doorkeeper.  I think doing it in Doorkeeper makes
most sense, as the special routing logic shouldn't apply to all
requests to the sessions controller, only those where they have come
from Doorkeeper deciding the user needs to authenticate.