If ?state has a ":", use first segment as JWT ID
This is a temporary work-around needed until we have an app with a
database handling the full auth flow on the www.gov.uk side.

The account-api will get around this problem by recording the redirect
URI against the state param on its end.  But as finder-frontend
doesn't have a database, it needs to pass the full thing in the state
param.