Add Webauthn Registration Feature
Adds the full feature allowing the user to register a new key.

Specific parts worth noting:

**Client side JS**

A lot of this needs to be handled client side using javascript,
here we've used a promise chain triggering closures that call functions
for each step.

There is some complexity here, mostly becuase the server-side wants
everything to be in base64Urls, and the browser API wants to recieve
everything as an ArrayBuffer. So we need to encode and decode our JSON.

There were oddities we encountered, for instance trying to pass an
ArrayBuffer straight back into JSON fails silently and leads to an empty
payload.

**Difficulty on classic GOV.UK stack bringing in NPM modules**

The complexity here could be outsourced to a NPM module. Github maintain
a well written clean module in typescript called
[Webauthn-json][webauthn-json].

Currently our GOV.UK Rails apps are still using sprokets for
compatibility with the GOV.UK publishing components. This prevented us
easily bringing these dependencies in.

Future attempts might do one of two things:

1. Keep using sprokets, but create a separate repository that takes the
github module, compiles it to plain old JS, then publishes it on NPM as
something like `gds-webauthn-json`. We could then bring that in with
node modules and sprokets.

2. Moderise our web apps to use webpacker. This is a bigger job that
would require modernising lots of JS across the platform. But this would
allow us to bring in dependencies such as bable, and add a transpile
step to our builds.

**Secure development environments**

We also encountered issues with getting access to the browser APIs in a
local dev setup. It turned out this because the `navigator.credentials`
API (core to this process) is a [secure-by-default
browser API][navigator-credentials].

The solution was to move our dev setup to a secure environment. We
explored:

- Using `mkcert` to generate local HTTPS certificates and apply them
using our docker setup or Puma. This could still be an option worth
investigating as it requires no change in URLs (leaving Plek
expectations unaffected), and brings our dev environment closer to a
12-factor app ideal of parity with production. In the end we decided to
go for a simplier route to unblock the issue more quickly.
- It was mentioned using a service that would route local dev though an
external web service that would expose it on an acutal web URL with a
real service SSL cert. However this seemed higher risk during development
and would pin us to using a new external service.
- Finally you can change URL to a reserved domain name that is pinned to
the loopback address, that the browser agrees to treat as secure. That's
what we went for here using `.localhost` in
[govuk-docker][docker-secure-env-pr]. It was quick, and worked, but
likely cannot be merged into GOV.UK docker as it breaks the expectations
of our Plek code that substitutes #{environment}.gov.uk for dev,
integration and staging.

**Difficulty feature testing**

We've left some notes in other PRs about the difficulties feature
testing this. Ideally we hoped to have one test that would cover both
our JS and ruby code from the perspective of a user (I register, now I
can see my key in security, and I can authenticate with it). We couldn't
find capybara methods that would hook into the browser to enable this.

Tests are left as notes to save some time for the next folks who try to
impliment this.

[docker-secure-env-pr]: https://github.com/alphagov/govuk-docker/pull/457
[webauthn-json]: https://github.com/github/webauthn-json
[navigator-credentials]: https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts
[secure-contexts]: https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>