Configure Webauthn to expect base64url encoding
We ran into some encoding expectation issues when trying to parse
external IDs during registration and authentication. The solution was to
switch from using `Base64.strict_encode` to `Base64.urlsafe_encode` with
the `padding: true` option.

Here we can additionally put this expectation explicitly into the
configuration of the app.

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>