Reorder the change password screen
This reorders the two fields on the Change password screen so that the
current password field is first and the new password field is second.

We've seen evidence that the current layout confuses both password
managers and some users. This should make it more friendly for both.

Trello card: https://trello.com/c/fcFsQuuD/611-reorder-the-change-password-screen