Move backlink before main
One of the issues highlighted in the Accounts DAC report was that screen
reader users reported confusion in encountering the back link in the
main content landmark.
The main landmark should not contain navigation elements such as the
back link. This moves the back link to just before the main.