Install Google Chrome on a separate later step
When building Docker images, the first step run is the the first that
changed in the Dockerfile. All steps before are not run again.

The installation of Google Chrome was mixed in the same step as general
updates and installation of postgresql-11 and redis-server.

This commit splits updates, installation of postgresql-11/redis-server,
and installation of Google Chrome, moving the latter to a later step so
that, if we need to change the Chrome version, the previous steps are
not run again and the build is faster.

Trello card: https://trello.com/c/icPgTrRp/660-investigate-concourse-pipeline-failure

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Dana Cotoran <Dana.Cotoran@digital.cabinet-office.gov.uk>