Pin Google Chrome to specific version
On 2021-03-02, [our Concourse pipeline started failing][0] with a
`Net::ReadTimeout: Net::ReadTimeout with #<TCPSocket:(closed)>` error.
The stack trace suggested that running `jasmine:ci` (originally included
in our default rake task) failed to start a server.

This happened with seemingly no code changes on our part, and we know
that Concourse is aggressively stateless so we investigated what could
have changed between the latest successful `run-quality-checks` job and
the first that failed. Finding no major changes there (different gem
versions, etc), we compared the `build-tests-image` jobs, specifically
the one present at the moment of the last successful test run and the
one present at the first failed test run. We realised that the only
major change was the version of Google Chrome used. Since `jasmine:ci`
runs depend on headless Chrome, this looked like the possible culprit
and not a coincidence.

We use the latest stable Chrome version, which is what Google recommends
but when the stable version changed from 88.0.4324.182-1 to
89.0.4389.72-1, this error started occurring.

We tried using the beta channel of Google Chrome but the issue doesn't
seem to be fixed yet, so we are pinpointing the Chrome version to the
latest one that we know works.

This will need to be revisited once we know the latest stable version
works and possibly again when we have a better strategy in place,
because [it happened before][1] and it's likely it will happen again.

As [suggested][2], treating Chrome as any other dependency (pinning it
to a specific version) is probably the way to go but we should have a
[process in place to update it regularly][3].

[0]: https://cd.gds-reliability.engineering/teams/govuk-tools/pipelines/govuk-account-manager-prototype/jobs/run-quality-checks/builds/1110
[1]: https://github.com/alphagov/publishing-e2e-tests/pull/336
[2]: https://github.com/alphagov/publishing-e2e-tests/pull/336#pullrequestreview-249971762
[3]: https://github.com/alphagov/publishing-e2e-tests/pull/336#issuecomment-502157397

Trello card: https://trello.com/c/icPgTrRp/660-investigate-concourse-pipeline-failure

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Dana Cotoran <Dana.Cotoran@digital.cabinet-office.gov.uk>