Fix feedback form showing an email error for the name field
The error_items helper takes input of the form

    [{ field: "name", text: "message" }, { field: "name", ...

So to extract the errors for a specific field we have to filter on the
`field` value.  And the conditional does that.

But we weren't doing that when constructing the error string, we were
filtering on the whole hash converted to a string, and since the error
message contained the word "name", it was being returned for the
"name" field.