Add type="tel" to telephone input fields
It was highlighted on our DAC report that the virtual keyboard which
appeared for number fields on mobile did not correspond with the
required input characters.
The virtual keyboard used for numerical/phone input should be the
numerical one. It would seem that our phone input fields did not have a
type attribute set, and would therefore default to using type="text".
This is why the alphabetical keyboard would appear, rather than the
correct numerical one.

This adds type="tel" to phone input fields, which will ensure that the
correct keyboard is displayed on mobile browsers.