Add summary error messages on views missing them
It was flagged as part of our DAC report that some pages on Accounts are
not using the validation error pattern suggested in the Design System:
design-system.service.gov.uk/patterns/validation/

Based on the DS recommendation, if the user's answers fail validation we
should show the error summary at the top of the page, and move keyboard
focus to it. Some of the pages that seemed to be missing this validation
pattern were the Sign in page, MFA code page, Edit phone page.

The purpose of this change is for accessibility and usability purposes –
all users should immediately be aware that a validation error has occured,
and should be able to navigate to the error fairly easily.