Add has_received_2021_survey to User table
In 2021 we've been sending a survey out to users who opt-in to give
feedback. We've wanted to know what they think about accoutns and what
they'd want to see it do.

We've been sending these in waves, starting with small pilot groups then
building up. In order to prevent sending the same person the survey
more than once we've been keeping a track of IDs using notes.

The next survery tranche is quite big, and we don't want a document full
of tons of IDs, so now is the right time to make this a boolean on the
table and flick it when we know we've tried to send the user the
survey.