Add a final safeguards to ensure we don't double send email
The Not clauses above this should exclude any duplicates between
groups... however when looking at the past cohort it seemed that instead
of the intented 500 emails sent we sent 498 because two users appeared
in two of the groups.

My concern is, if these are not being filtered out we might send a small
number of users two emails which would be annoying. So let's enhance
this by updating the database as we go.

See here for more details[1]

[1]: https://docs.google.com/document/d/1Ojg57BOtFxXa_3glvjApHxISnwgRMvosplXC4KNrFhk/edit?disco=AAAAL3PadlM