Send the survey in chunks & sleep between each
This is to avoid hitting the Notify rate limit, which we did
yesterday.  I've gone for 2500, rather than 3000, so that we have
wiggle room for normal emails to still go out at the same time.

We may not end up using this task again.  But by putting the code here
now it'll serve to remind us the next time we want to do something
like this.