Add a LevelOfAuthentication module
We are going to need some calculations that may need to be shared, but
mostly here I just want them not clutting up initalizers.

The first challenge, we get a list of scopes, I cannot guarentee order
and they might be multiple, so lets grab them, sort the by the final
integer and give a way of grabbing the highest level.