Add default maximum level as level1
Given a range of levels we want to select the highest.

But what if we are passed an empty array?

In this case it is best that we define in advance the highest level of
auth we know about, and force the user to auth to that level for safety.

In reality this should only occur temporarily of a dev forgot to do
something. So it is a safety net.