Fix Change email page bug and add tests
On the Change your email page, there is a bit of content that notifies
the user what their account's email address currently is.
The content reads: The email address currently stored in your GOV.UK
account is: something@actualemail.com

What is happening is that if the user submits the form with the wrong
information, so that an error appears, the content above changes to
reflect (incorrectly) what the user had typed into the email input prior
to the error. This could be misleading because people might think that
their email has been changed, when it fact it has stayed the same.