Use heading_level param instead of is_page_heading
The account manager uses the radio component with the is_page_heading
option. The radio component is being refactored to remove the
is_page_heading parameter in favour of using heading_level and/or
heading_size to achieve the same result.
In this case, is_page_heading: true can simply be swapped for
heading_level: 1 without any visual changes.