Always show the Transition Checker service card
We don't need to check if the user has used the Transition Checker,
because you can't sign up without going via it.

We are going to launch another experiment, the save-a-page experiment,
but for that we'll be designing and building a separate "your account"
style page on GOV.UK itself, and retiring this page.

Removing this conditionality here will make this page faster to load,
reduce requests to the Attribute Service, and make it easier to
migrate the `transition_checker_state` attribute from the Attribute
Service and into the Account API (though that's unlikely to get done
before we retire this page anyway).