Switch on new Design System link styles
The new DS link styles are being rolled out across GOVUK. We've
already applied the new styles to the individual components in the
components gem, so the next step is to switch the flag on in the
frontend apps themselves.
This sets $govuk-new-link-styles to true, which enables the new styles
on the account manager prototype.