Move flash alerts to manage page
We're moving the "your account" page onto GOV.UK, and don't want to
have to figure out cross-app flash messages just yet.

I've removed the `set_flash_message_for_update`, as that was only used
to show the flash for when you've changed your email address; but we
have a separate banner for that.

I've also changed the view to report an error to Sentry if it gets a
flash message, which will help us track down any others.