Remove JWT auth flow
We originally implemented this so we could still persist the user's
attributes even if they *don't* click the link on the registration
page to return to GOV.UK.  This was, with hindsight, a mistake: it's a
non-standard OAuth flow, won't be supported by Digital Identity, and
prevented us from keeping in mind a clear separation between the
authentication journey and the service.

We have now stopped sending JWTs from the Transition Checker, which
was the only thing to use this functionality.  So now we can remove
it.

See also https://github.com/alphagov/finder-frontend/pull/2575