Remove the Transition Checker email choice page
We've decided that it makes most sense to subscribe the user to
notifications by default.  Given that the user first has to choose
between the account and email notifications, it's a bit strange to ask
them if they want email notifications again.

A corresponding change will need to be made in finder-frontend, as
that prompts the user to sign up for notifications when saving their
results if they don't already have a subscription.