Change "return" link on post-registration page to a button
I had to introduce a controller which just performs a redirect to the
previous URL, which is a bit awkward.  It can't just be a GET form
with action `previous_url` because `previous_url` has a querystring
and GET forms can't have a querystring in the action.

I also changed the `continue_to_service` text to not be generic.  We
only have one service using the account manager: GOV.UK.  When we end
up in Digital Auth land we'll still only be one app, so we should
start to get used to only having one piece of text.