Remove email subscription management
The account-api has been changed to no longer create subscriptions in
this app, so we can remove all the code for managing them here.  I've
also revoked the email-alert-api bearer token.

I've left the EmailSubscription model in place, rather than drop the
table, so that we still have the current state if we need it, but it
won't change from now on (other than user deletions).