Change layout of confirmation page when signing up via a service
We've got new designs for this page in the case where a user signs up
through a service.  Since the same template is used for both
registration confirmation and change-of-email confirmation (whose
bright idea was that? oh right, me) there's a bit of conditional logic
in it now to render the right things.

We don't have a new design for the case where a user signs up by going
to `/new-account` directly (not via a service), so I've kept that
close to old design.