Remove account-api data when deleting the account
We are in the process of reducing the amount the Account manager holds
and has to do in preparation for this element to become a commodity
plug-and-play service we can hand over to another GDS team.

During this transition we have two stores of users data, some in the
account manager and others in the account-api. This PR ensures the two
are in lock step when a user deletes their account, so we do not
continue to hold data in the account-api for users who have deleted
their account through the manager prototype