Remove conditionality from confirmation screen
In govuk-account-manager-prototype#911 we added conditionality that
varied the confirmation page based on the user's origin. This was an
attempt to provide some content to users that were specific to their
jouernys.

However this has a few drawbacks.

1. The account manager can only really tell if the user arrived via the
auth path, or did not. This is a reasonable proxy for "signed up from
our only service" or "signed up directly", but does not really prove
they came from the brexit checker. This has led to a few odd journeys
where the user comes to the sign up journey via gov.uk/sign-in, which
preserves previous_url parameters. This also will not work when we
immimently add a second service.

2. It also implies a level of knowlege / access to GOV.UK storage
(brexit checker attribtutes) that we cannot expect a Digital Identity
product to have. It is therefore a re-entangling of the checker and the
auth journey, whilst we're trying to separate the two.

As a result we've chosen here to opt for a single generic completion
page that sends the user back to the previous_url