Upgrade level0 users to level1, if they have MFA configured
Users can log in at level0, but account-api enforces levels of
authentication.  This means that we need to be able to reauthenticate
a user at level1 in this app.

We only have placeholder content for the MFA set-up journey, but
currently users can only create a level1 account, so this works:

- If the user has MFA, and is authenticated to too low a level,
upgrade them.

- If the user does not have MFA, and is authenticated to too low a
level, do not upgrade them.

This will result in broken journeys for users without MFA, but there
are no such users.