Persist subject identifiers to the database
When we migrate to DI, all the internal database identifiers will
change, and so will the subject identifiers.  DI are persisting the
old subject identifiers as a custom claim, so that we can match up
data if we need to (eg) examine old audit logs or suchlike.  Storing
the subject identifiers here as well makes it possible to do that just
with database queries, which will be much faster than running Ruby
code.