Stop users creating new accounts
We are migrating to use Digital Identity's sign in service, and while
this is happening we need to stop users from creating new accounts.

After the migration, the account manager prototype will be retired, so
we can safely delete the registration tests.

Also redirect the pages in the account creation flow to the 'start' page.
We've stopped people from creating a new account, so they shouldn't be
accessing these pages. Redirect them to account creation start page so
noone can register or update an account during migration.

I put the redirects in the `routes` file rather than each controller
method so it's easier to see what's happening to all the routes together.