Remove password reset form
We've already removed the route this form points to, so anyone going to
the password reset page from an earlier link would get a 5xx error
because the path `user_password_path` is no longer defined.

Remove the form from the page and replace it with the 'You cannot update
this information at the moment' warning that we use on the other removed
forms.