Return a 404 for all unknown paths when redirecting
For some paths in production we seem to be serving a 500 response
instead of a 404.  I'm not really sure why, and this wouldn't really
be a problem, but we still have the alerting for too-high 5xx rate,
and traffic to this app has died down.  So ideally we won't be
serving *any* 5xx responses, or we risk calling someone until that
alert gets removed.

I've raised a PR to remove the alert, but RE won't get to it until
tomorrow.  To avoid the risk of calling someone overnight, let's add
some special handling for missing files.